Potential fix for code scanning alert no. 44: Multiplication result converted to larger type

+++ src/audio/SDL_audioqueue.c
@@ -537,7 +537,7 @@ const Uint8 *SDL_ReadFromAudioQueue(SDL_AudioQueue *queue,
     int src_channels = track->spec.channels;
     const int *src_map = track->chmap;
 
-    size_t src_frame_size = SDL_AUDIO_BYTESIZE(src_format) * src_channels;
+    size_t src_frame_size = ((size_t) SDL_AUDIO_BYTESIZE(src_format)) * src_channels;
     size_t dst_frame_size = ((size_t) SDL_AUDIO_BYTESIZE(dst_format)) * dst_channels;
 
     size_t src_past_bytes = past_frames * src_frame_size;
